Module foundry_common_fmt::dynamic

Structs

DynValueDisplay 🔒

Wrapper that implements Display for a [DynSolValue].

DynValueFormatter 🔒

[DynSolValue] formatter.

Functions

format_token

Pretty-prints the given value into a string suitable for user output.

format_token_raw

Pretty-prints the given value into a string suitable for re-parsing as values later.

format_tokens

Pretty-prints a slice of tokens using format_token.

format_tokens_raw

Pretty-prints a slice of tokens using format_token_raw.

parse_tokens

Parses string input as Token against the expected ParamType

print_tokens

Prints slice of tokens using format_tokens or format_tokens_raw depending on json parameter.