foundry_common_fmt

Module dynamic

Structs

DynValueDisplay 🔒

Wrapper that implements Display for a [DynSolValue].

DynValueFormatter 🔒

[DynSolValue] formatter.

Functions

format_token

Pretty-prints the given value into a string suitable for user output.

format_token_raw

Pretty-prints the given value into a string suitable for re-parsing as values later.

format_tokens

Pretty-prints a slice of tokens using format_token.

format_tokens_raw

Pretty-prints a slice of tokens using format_token_raw.

parse_tokens

Parses string input as Token against the expected ParamType

serialize_value_as_json

Serializes given [DynSolValue] into a serde_json::Value.